+++ sky/engine/core/painting/Matrix.h
@@ -6,13 +6,13 @@
 #define SKY_ENGINE_CORE_PAINTING_MATRIX_H_
 
 #include "sky/engine/bindings/exception_state.h"
-#include "sky/engine/tonic/float32_list.h"
+#include "sky/engine/tonic/float64_list.h"
 #include "third_party/skia/include/core/SkMatrix.h"
 
 namespace blink {
 
-SkMatrix toSkMatrix(const Float32List& matrix4, ExceptionState& es);
-Float32List toMatrix4(const SkMatrix& sk_matrix);
+SkMatrix toSkMatrix(const Float64List& matrix4, ExceptionState& es);
+Float64List toMatrix4(const SkMatrix& sk_matrix);
 
 }  // namespace blink
